+++ javaagent-tooling/src/test/groovy/io/opentelemetry/javaagent/tooling/muzzle/InstrumentationClassPredicateTest.groovy
@@ -15,10 +15,9 @@ class InstrumentationClassPredicateTest extends Specification {
     InstrumentationClassPredicate.isInstrumentationClass(className)
 
     where:
-    desc                            | className
-    "auto instrumentation class"    | "io.opentelemetry.javaagent.instrumentation.some_instrumentation.Advice"
-    "javaagent-tooling class"       | "io.opentelemetry.javaagent.tooling.Constants"
-    "library instrumentation class" | "io.opentelemetry.instrumentation.LibraryClass"
+    desc                              | className
+    "javaagent instrumentation class" | "io.opentelemetry.javaagent.instrumentation.some_instrumentation.Advice"
+    "library instrumentation class"   | "io.opentelemetry.instrumentation.LibraryClass"
   }
 
   @Unroll
@@ -29,6 +28,7 @@ class InstrumentationClassPredicateTest extends Specification {
     where:
     desc                        | className
     "Java SDK class"            | "java.util.ArrayList"
+    "javaagent-tooling class"   | "io.opentelemetry.javaagent.tooling.Constants"
     "instrumentation-api class" | "io.opentelemetry.instrumentation.api.InstrumentationVersion"
     "javaagent-api class"       | "io.opentelemetry.javaagent.instrumentation.api.ContextStore"
   }
